requirements.txt
alembic==1.11.3
blinker==1.6.2
click==8.1.7
flask==2.3.3
Flask-JWT-Extended==4.5.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
importlib-metadata==6.8.0
importlib-resources==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
PyJWT==2.8.0
PyMySQL==1.1.0
python-dotenv==1.0.0
SQLAlchemy==2.0.20
typing-extensions==4.7.1
Werkzeug==2.3.7
zipp==3.16.2
mysqlclient
Flask-Migrate
